requirements.txt
pyautogen>=0.10.0
python-dotenv>=1.0.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.0.0
